Refactored tracing as well
