+++ app/routes.py
@@ -4,14 +4,12 @@ from fastapi.responses import JSONResponse
 from typing import Optional, Dict
 import uuid
 import os
-from contextlib import asynccontextmanager
-from .session import MCPLocalSessionTask, mcp_session, try_get_session_id, session_id
-from .session_manager import session_manager
-from .models import RunToolsResult
+
+from app.utils.traced_requests import traced_request
+from app.session import MCPLocalSessionTask, try_get_session_id, session_id
+from app.session_manager import mcp_session_context, session_manager
 from .mcp_server import get_server_params
-from .oauth.decorator import decorate_args_with_oauth_token
 from .oauth.user_info import get_data_access_manager
-from fnmatch import fnmatch
 from opentelemetry import trace
 from .oauth.token_exchange import UserLoggedOutException
 from .utils import mask_token
@@ -30,112 +28,8 @@ EXCLUDE_TOOLS = [t for t in os.environ.get("EXCLUDE_TOOLS", "").split(",") if t]
 
 tracer = trace.get_tracer(__name__)
 
-
-def matches_pattern(value, pattern):
-    """Check if a value matches a glob-like pattern."""
-    return fnmatch(value, pattern)
-
-
-def map_tools(tools):
-    """Map tools with INCLUDE_TOOLS and EXCLUDE_TOOLS filters applied."""
-
-    filtered_tools = []
-    for tool in tools.tools:
-        include_match = any(
-            matches_pattern(tool.name, pattern) for pattern in INCLUDE_TOOLS if pattern
-        )
-        exclude_match = any(
-            matches_pattern(tool.name, pattern) for pattern in EXCLUDE_TOOLS if pattern
-        )
-        logger.debug(
-            f"[Tools] Filter check -> Tool: {tool.name}, Include Match: {include_match} - {INCLUDE_TOOLS}, Exclude Match: {exclude_match} - {EXCLUDE_TOOLS}"
-        )
-        if INCLUDE_TOOLS and any(INCLUDE_TOOLS) and not include_match:
-            continue
-        if EXCLUDE_TOOLS and any(EXCLUDE_TOOLS) and exclude_match:
-            continue
-
-        filtered_tools.append(
-            {
-                "name": tool.name,
-                "title": tool.title,
-                "description": tool.description,
-                "inputSchema": tool.inputSchema,
-                "outputSchema": tool.outputSchema,
-                "annotations": tool.annotations,
-                "meta": tool.meta,
-                "url": f"{MCP_BASE_PATH}/tools/{tool.name}",
-            }
-        )
-
-    return filtered_tools
-
-
 if MCP_BASE_PATH:
     router.prefix = MCP_BASE_PATH
-
-
-@asynccontextmanager
-async def mcp_session_context(x_inxm_mcp_session: Optional[str], access_token: Optional[str], group: Optional[str]):
-    """Yield a delegate with unified list_tools() and call_tool() across sessionful and sessionless modes."""
-    # Sessionless path: validate group access (if present) and open a transient MCP session
-    if x_inxm_mcp_session is None:
-        if group and access_token:
-            data_manager = get_data_access_manager()
-            try:
-                data_manager.resolve_data_resource(access_token, group)
-            except PermissionError as e:
-                logger.warning(f"Group access denied: {str(e)}")
-                raise HTTPException(status_code=403, detail=str(e))
-
-        async with mcp_session(get_server_params(access_token, group)) as session:
-            class SessionDelegate:
-                async def list_tools(self):
-                    tools = await session.list_tools()
-                    return map_tools(tools)
-
-                async def call_tool(self, tool_name: str, args: Optional[Dict], access_token_inner: Optional[str]):
-                    tools = await session.list_tools()
-                    decorated_args = await decorate_args_with_oauth_token(
-                        tools, tool_name, args, access_token_inner
-                    )
-                    result = await session.call_tool(tool_name, decorated_args)
-                    return RunToolsResult(result)
-
-            yield SessionDelegate()
-        return
-
-    # Sessionful path: reuse existing task, but surface a common delegate API
-    mcp_task = sessions.get(x_inxm_mcp_session)
-    if not mcp_task:
-        logger.warning(
-            mask_token(
-                f"[MCP] Session not found: {x_inxm_mcp_session}",
-                x_inxm_mcp_session,
-            )
-        )
-        raise HTTPException(status_code=404, detail="Session not found. It might have expired, please start a new.")
-
-    class TaskDelegate:
-        async def list_tools(self):
-            tools = await mcp_task.request("list_tools")
-            return map_tools(tools)
-
-        async def call_tool(self, tool_name: str, args: Optional[Dict], access_token_inner: Optional[str]):
-            tools = await mcp_task.request("list_tools")
-            decorated_args = await decorate_args_with_oauth_token(
-                tools, tool_name, args, access_token_inner
-            )
-            result = await mcp_task.request(
-                {"action": "run_tool", "tool_name": tool_name, "args": decorated_args}
-            )
-            return RunToolsResult(result)
-
-    try:
-        yield TaskDelegate()
-    finally:
-        # No-op: persistent session remains managed elsewhere
-        pass
 
 
 @router.get("/tools")
@@ -148,24 +42,20 @@ async def list_tools(
     ),
 ):
     try:
-        with tracer.start_as_current_span("list_tools") as span:
-            x_inxm_mcp_session = session_id(
-                try_get_session_id(
-                    x_inxm_mcp_session_header, x_inxm_mcp_session_cookie
-                ),
-                access_token,
-            )
-            if x_inxm_mcp_session:
-                span.set_attribute("session.id", x_inxm_mcp_session)
-            if group:
-                span.set_attribute("session.group", group)
-            logger.info(
-                mask_token(
-                    f"[Tools] Listing tools. Session: {x_inxm_mcp_session}, Group: {group}",
-                    x_inxm_mcp_session,
-                )
-            )
-            async with mcp_session_context(x_inxm_mcp_session, access_token, group) as session:
+        x_inxm_mcp_session = session_id(
+            try_get_session_id(x_inxm_mcp_session_header, x_inxm_mcp_session_cookie),
+            access_token,
+        )
+        with traced_request(
+            tracer,
+            operation="list_tools",
+            session_value=x_inxm_mcp_session,
+            group=group,
+            start_message=f"[Tools] Listing tools. Session: {x_inxm_mcp_session}, Group: {group}",
+        ):
+            async with mcp_session_context(
+                sessions, x_inxm_mcp_session, access_token, group
+            ) as session:
                 result = await session.list_tools()
                 logger.debug(
                     mask_token(
@@ -192,31 +82,28 @@ async def run_tool(
     ),
 ):
     try:
-        with tracer.start_as_current_span(
-            "run_tool", attributes={"tool.name": tool_name}
-        ) as span:
-            x_inxm_mcp_session = session_id(
-                try_get_session_id(
-                    x_inxm_mcp_session_header,
-                    x_inxm_mcp_session_cookie,
-                    args.get("inxm-session", None) if args else None,
-                ),
-                access_token,
-            )
-            if x_inxm_mcp_session:
-                span.set_attribute("session.id", x_inxm_mcp_session)
-            if group:
-                span.set_attribute("session.group", group)
-            if args and "inxm-session" in args:
-                args = dict(args)
-                args.pop("inxm-session")
-            logger.info(
-                mask_token(
-                    f"[Tool-Call] Tool call: {tool_name}, Session: {x_inxm_mcp_session}, Group: {group}, Args: {args}",
-                    x_inxm_mcp_session,
-                )
-            )
-            async with mcp_session_context(x_inxm_mcp_session, access_token, group) as session:
+        x_inxm_mcp_session = session_id(
+            try_get_session_id(
+                x_inxm_mcp_session_header,
+                x_inxm_mcp_session_cookie,
+                args.get("inxm-session", None) if args else None,
+            ),
+            access_token,
+        )
+        if args and "inxm-session" in args:
+            args = dict(args)
+            args.pop("inxm-session")
+        with traced_request(
+            tracer=tracer,
+            operation="run_tool",
+            session_value=x_inxm_mcp_session,
+            group=group,
+            start_message=f"[Tool-Call] Tool call: {tool_name}, Session: {x_inxm_mcp_session}, Group: {group}, Args: {args}",
+            extra_attrs={"tool.name": tool_name},
+        ):
+            async with mcp_session_context(
+                sessions, x_inxm_mcp_session, access_token, group
+            ) as session:
                 result = await session.call_tool(tool_name, args, access_token)
 
         logger.info(f"[Tool-Call] Tool {tool_name} called. Result: {result}")
