+++ app/session_manager/__init__.py
@@ -1,3 +1,4 @@
 from .session_manager import session_manager
+from .session_context import mcp_session_context
 
-__all__ = ["session_manager"]
+__all__ = ["session_manager", "mcp_session_context"]
